Move packing logic
